[fix] refactor builds and passes tests
