+++ src/index.ts
@@ -1,18 +1,15 @@
-// workaround for https://github.com/parcel-bundler/parcel/issues/5911
 import * as _MoreMath from './more-math'
-
-export { xin, observe, unobserve, touch, observerShouldBeRemoved } from './xin'
-export { xinProxy } from './xin-proxy'
-export { hotReload } from './hot-reload'
-export { throttle, debounce } from './throttle'
-export { Component } from './component'
-export { elements, makeComponent } from './elements'
-export { settings } from './settings'
 export { bind, on } from './bind'
 export { bindings } from './bindings'
-export { getListItem, xinPath, xinValue } from './metadata'
-export { vars, initVars, css, darkMode } from './css'
+export { vars, initVars, css, darkMode, XinStyleSheet, XinStyleMap, XinStyleRule } from './css'
 export { Color } from './color'
+export { Component } from './component'
+export { elements, makeComponent, ElementsProxy } from './elements'
+export { hotReload } from './hot-reload'
+export { getListItem, xinPath, xinValue } from './metadata'
 export const MoreMath = _MoreMath
-
+export { settings } from './settings'
+export { throttle, debounce } from './throttle'
+export { xin, observe, unobserve, touch, observerShouldBeRemoved } from './xin'
 export * from './xin-types'
+export { xinProxy } from './xin-proxy'
